requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
